pyproject.toml
[project]
name = "lajeh-api"
version = "0.1.0"
description = "Chatbot de follow-up"
authors = [
    {name = "BrunaCardoso7",email = "obitadrawing@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12.*"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alembic (>=1.14.0,<2.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"